requirements.txt
fastapi==0.70.0
requests==2.26.0
uvicorn==0.16.0
python-multipart==0.0.5
python-dotenv==0.17.0
Jinja2==2.11.1
markupsafe==2.0.1
